Refactor defaults to use config-driven paths
